app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10
pika==1.3.2
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
bcrypt==4.3.0
